moved a couple of things around and added a  flask file
